requirements.txt
# Core FastAPI stack
fastapi[all]
uvicorn[standard]==0.35.0
starlette==0.47.3
werkzeug==3.1.3
MarkupSafe==3.0.2

# MongoDB async driver
motor==3.7.1
pymongo==4.14.1

# Data validation and utilities
pydantic==2.11.7
pydantic-core==2.33.2
email-validator==2.3.0
python-dateutil==2.9.0.post0

# Security & authentication
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
rsa==4.9.1
pyasn1==0.6.1

# Environment & config
python-dotenv==1.1.1
pyyaml==6.0.2

# HTTP clients / API utilities
python-http-client==3.3.7
sendgrid==6.12.4

# Async / networking
anyio==4.10.0
websockets==15.0.1
# uvloop==0.21.0
watchfiles==1.1.0
typing-extensions==4.15.0
typing-inspection==0.4.1
sniffio==1.3.1

# Optional utilities
click==8.2.1
six==1.17.0
ecdsa==0.19.1
pytz
pandas
requests
resend
# deactive schedulers 
apscheduler==3.10.4

#opening an xml doc
openpyxl==3.1.5
